実行方法変更。

+++ example/example.sh
@@ -1,5 +1,10 @@
-#!/bin/bash
+#!/bin/bash -e
+
+cd $(dirname $0)
+
+if [ -d output ]; then
+  rm -r output
+fi
 
 hadoop jar WordCountMR-1.0-SNAPSHOT.jar input output
 cat output/*
-rm -r output
